rfct: Replace login.defs regex with a custom parser
The format is so simple that a custom parser would be much safer and
faster than using regex. The original also should have used `^` instead
of `\n`.

This is one of the steps towards removing regex as a dependecy for this
project and replacing it with parsers for the specific file formats.